# src/Web/Client/app/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Vocab.Web.Api",
    "version": "1.0"
  },
  "paths": {
    "/api/hello": {
      "get": {
        "tags": [
          "Vocab.Web.Api"
        ],
        "operationId": "SayHello",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/String<>f__AnonymousType0"
                }
              }
            }
          }
        }
      }
    },
    "/api/languages": {
      "post": {
        "tags": [
          "Vocab.Web.Api"
        ],
        "operationId": "CreateLanguage",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateLanguageCommand"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "get": {
        "tags": [
          "Vocab.Web.Api"
        ],
        "operationId": "GetAllLanguages",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/languages/{id}": {
      "get": {
        "tags": [
          "Vocab.Web.Api"
        ],
        "operationId": "GetLanguageById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "tags": [
          "Vocab.Web.Api"
        ],
        "operationId": "UpdateLanguage",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateLanguageCommand"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "CreateLanguageCommand": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "String<>f__AnonymousType0": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "UpdateLanguageCommand": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    }
  }
}